Add CartesianPlanning method & cleanup code
This commit adds the CartesianPlanning method which takes
a pose input argument and moves the end-effector after
planning.

Additionally the InverseKinematics method - getIK(), now works
and the create_pose method is implemented inside the other
methods rather than doing the call from the main() execution
section.

At last the naming convention is now uniform throughout the
code for easier navigation. Ex. goal_pose -> goalPose

TODO: Seemingly there is a bug in the CartesianPlanning method
where the first run doesn't manage to create a plan and the
status is: ABORT. However a second run in sequence always manages
to create a plan and execute the movement.