Add GetJacobianMatrix method
Added the GetJacobianMatrix functionality which returns a 6xn
array. Additionaly added a getEeVelocityVector method which is
still mocked and has to be tested before implementation.
It is supposed to calculate the linear and angual velocity of the
end effector which would be used to calculate joint velocities
used for movement execution.